Use a Poisson distribution to model rate of bug reports on Google code and Github.
Summary:
Further, on Github, ignore reports from the same user that happen very
quickly because they are likely to not be real bugs.
We now only report if the probability is very very high (at
least 0.99 or 0.997 on google code and github, respectively) that the
report rate is abnormally elevated.

Test Plan: Ran locally and verified functionality in HipChat Tests room.

Reviewers: jascha, csilvers

Reviewed By: csilvers

CC: csilvers

Differential Revision: http://phabricator.khanacademy.org/D605