Use alertlib (now a submodule) to send to hipchat.
Summary:
This is mostly so we can re-use the existing secrets-management we
have with alertlib, rather than having to set up secrets again for
beep-boop.  But it's also for consistency with other apps on toby.

Test Plan:
Ran
   env PYTHONPATH=alertlib:$HOME/khan/webapp ./uservoice_reports.py
and saw it fail on uservoice.cfg, meaning that at least it compiles
cleanly. :-)

Reviewers: alpert

Reviewed By: alpert

Subscribers: chris

Differential Revision: http://phabricator.khanacademy.org/D11440