Move from uservoice over to zendesk.
Summary:
The zendesk API is a bit annoying, because they rate limit pretty
aggressively.  But overall, it's pretty straightforward.

I still have to update the cronjob (in Khan/aws-config) and update the
password file on toby before this can go live.

Test Plan:
Ran manually with success.
Ran again manually with success (it slept for about a minute).
Fingers crossed that it will actually alert properly...

Reviewers: alpert

Reviewed By: alpert

Subscribers: nathan

Differential Revision: http://phabricator.khanacademy.org/D12961