Added per-exercise issue monitoring for JIRA
Summary:
This diff adds per-exercise error monitoring for JIRA to our suite
of monitoring tools. See the [design doc](https://docs.google.com/a/khanacademy.org/document/d/12Hkg0vSO3whawOveyTwBUh0UyyT7uEkh2MqAHyxZWWU/edit?usp=sharing)
for specific project goals.

Most of the changes are in jira_reporting.py, the core logic of which follows
zendesk_reporting.py. The JIRA API returns a maximum of 1000 issues, so given
a timeframe, we ask for the first 1000 issues beggining at the start time, track
the most recent issue returned, and keep requesting until we've found all
relevant issues. Error rates are tracked on a per-exercise basis and stored as a
map from exercise name to number of errors seen.

As with the Zendesk module, on first run, we grab as much historical data as we can
(100 days, for now--much more than for Zendesk, since JIRA is per-exercise).
On future runs, we check any issues created since our last run and report back
if an exercise has an abnormally high error rate. (Allegedly) broken exercises
are posted on HipChat (in the jira-alertlib-test room for now, but eventually in the
Content Tools room, once this has been reviewed).

Test Plan:
- Pull the repo.
- Make sure you have alertlib setup properly.
- Run `python jira_reporting.py` to grab historical data. The first call can take a while (5 minutes, on my machine). If you care less about accuracy, considering changing num_days_in_past to something more manageable, like 7.
- The module should print a list of error counts for each exercise, like so:
    ```
    2015-01-20 16:07:38 EST] halves-and-fourths TOTAL 0/0s; 1421183400.0-: 9/604525.0s; m=0.000 p=1.000
    ```
    - Verify that the time elapsed roughly matches `num_days_in_past * 24 * 60 * 60`, e.g., above, I set `num_days_in_past = 7` and `7 * 24 * 60 * 60 ~= 604525`.
    - Verify that each exercise has a reasonable number of issues, e.g., above, we see 9 issues over the past week.
- Apply the diff [here](https://phabricator.khanacademy.org/differential/diff/46553/), which changes the query window to a period of known exercise breakage and prints, rather than logging to HipChat.
    - Verify that known broken exercises are reported as spiking, such as: `congruent_angles`, `properties-of-rigid-transformations`, and `scientific_notation_intuition`.
- Alternatively, to see exercises that have spiked recently:
    - Delete the `jira` file, if you've already run.
    - Change `now = int(time.time())` on line 141 to something like `now = int(time.time()) - 24 * 60 * 60` (to only scrape historical data up until today).
    - Run `python jira_reporting.py` to get historical data.
    - Change line 141 back to the original.
    - Run `python jira_reporting.py` again to see if any exercises have spiked today.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: eater, emily

Differential Revision: https://phabricator.khanacademy.org/D15800