Track starting time per-exercise and introduce reporting threshold
Summary:
Previously, we assumed that the first issue for every exercise was
created at the same time. This introduced a bias in which newer exercises
(i.e., those that were created within 100 days) would seem to have lower
historical rates for issue reporting and thus get flagged too frequently. For
example, an exercise created last week might appear to have only one issue
created in the past 100 days, when in reality, that should be treated as one
issue in the past week. Now, we track the time-of-first-issue for each exercise.

Additionally, we only report exercises if we see > 3 issues in the window.
Based on some trial runs, this seemed like a good value (some exercises were
being flagged for just a single issue in an hour, which was a bit excessive).

Test Plan:
- Run `python jira_reports.py` and wait for the historical data to aggregate.
    - Verify that the third-to-last value varies a bit for each exercise.
- Run `python jira_reports.py` an hour later and check the results.
- For a more elaborate test plan, you can recreate the test plan from D15800.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: eater

Differential Revision: https://phabricator.khanacademy.org/D16056