add slack support to beep boop
Summary: also adjusts for both hipchat and slack to not post to dead rooms.  dead stuff is going to support channel for now, can adjust later if requested.

Test Plan: There is no real great way to test this outside of production currently so instead of shaving a yak my plan is to just deploy it early and monitor it and see what happens, comparing output between hipchat and slack over the next few days and making adjustments

Reviewers: ethan, benjaminpollack

Reviewed By: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D20692