point jira reports at #content, indicate github reports no longer active
Summary:
Point the jira reports at #content based upon feedback from @eater.

The github reports are no longer being used in production.  It is
unclear when they were originally removed from production (it seems to
pre-date toby’s crontab being in version control) but add a comment to
the file to avoid anyone in the future assuming it’s running and
refactoring it.

Test Plan:
wait until we see an alert, if we receive negative feedback
adjust destination.

Reviewers: eater

Reviewed By: eater

Subscribers: eater, benjaminpollack, ethan

Differential Revision: https://phabricator.khanacademy.org/D20721