Update beep-boop to send to alerta regardless of error state
Summary:
In order to automatically close alerts in the Alerta dashboard, cronjobs must be setup to ping alerta when zendesk ticket rates are back to normal after an elevated state. This is achieved by sending an alert with the same args (save severity which gets translated to ‘cleared’ in the alerta script).

Also, updated alertlib submodule to reflect new severity ratings.

Test Plan: Tested locally

Reviewers: nabil, kevindangoor

Reviewed By: nabil

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D38248