Update alertlib and make alerta resolution explicit
Summary:
https://app.asana.com/0/412860343989559/427743112677951

We want to update code that uses alertlib to resolve
Alerta alerts to explictly indicate that so that we
can make resolution of info alerta not the default.

This diff updates alertlib and makes the zendesk
alert resolution explicit.

Test Plan: Fingers crossed. Shouldn't be a disaster if it turns out that there's a problem.

Reviewers: jacqueline, nabil

Reviewed By: jacqueline

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D39368