Make beep-boop treat evenings like weekdays
Summary:
beep-boop has been pretty noisy on weekday mornings of late.  This seems to be
in part because the "weekday" average ticket rate is artificially low since it
includes evenings.  Instead, to be a little smarter but keep things simple, we
now consider a hardcoded range of overnight hours to also be equivalent to
weekends.  This should cause the threshold for weekdays to rise a bit (and give
us a bit more off-hours data).  It's not clever but hopefully it'll be good
enough.

Test Plan:
deploy it, keep an eye on the beep-boop logs and make sure it
doesn't whine too much (which should be easy since I'm dev-support)

Reviewers: jacqueline, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D39513