Make sure mean is always a float.
We were getting this error:
```
 Traceback (most recent call last):
  File "/home/ubuntu/beep-boop/zendesk_reports.py", line 285, in <module>
    main()
  File "/home/ubuntu/beep-boop/zendesk_reports.py", line 249, in main
    time_this_period)
  File "/home/ubuntu/beep-boop/util.py", line 20, in probability
    return (mean, poisson_cdf(errors_this_period - 1, mean))
  File "/home/ubuntu/beep-boop/util.py", line 40, in poisson_cdf
    p = (-mean).exp()
AttributeError: 'int' object has no attribute 'exp'
```

This wouldn't normally happen, but there was a manual edit to the
beep-boop json on Friday that turned some floats into an int (see
https://app.asana.com/0/31965416896056/427370269499201/f).

We might as well be proactive and make *sure* `mean` is a float.

Auditors: ragini

Test plan:
Fingers crossed!