Link to #zendesk-tickets instead of zendesk in beep-boop message
The slack feed is a better place to look because you don't have to
figure out the right filter (e.g. if some of the tickets are already
closed) and it's easier anyway.  (Even more so in the Pager Parrot
version, which gets truncated, I think by PagerDuty, breaking the url.)

Test plan: crossed fingers

Auditors: jacqueline