Also catch EOFErrors when reading old data pickles
Summary:

Something last week caused us to generate some pickles of old data that
give EOFErrors when trying to read them. We already have code to handle
failure when reading pickles that just returns some default values, but
it doesn't handle this particular error. This commit changes it so that
we catch `EOFError` so that we can deal with corrupted pickles in
addition to missing ones.

Test Plan:
- `env PYTHONPATH=$HOME/alertlib_secret /usr/local/bin/cron_with_output -c2 '/usr/bin/timeout 55m env PYTHONPATH=$HOME/beep-boop/alertlib:$HOME/alertlib_secret python $HOME/beep-boop/zendesk_reports.py' $HOME/logs/beep-boop-zendesk.log`
- verify the command completes successfully

Auditors: drosile