Raise threshold for paging
Summary:
Recently we had a beep boop alert false positive that triggered
pagerduty.

I reviewed beep boop alerts for this year:

  date, probabity, num tickets, false or true alert
- feb 15, 1.0000, 41, true
- feb 15, 0.9999, 6, false
- feb 11, 0.9999, 6, false
- feb 06, 0.9991, 9, false (no page)
- feb 05, 0.9991, 9, true (no page)
- feb 05, 0.9991, 9, true (no page)
- feb 05, 0.9992, 9, true (no page)
- feb 03, 1.0000, 7, true
- feb 02, 0.9993, 5, true (no page)
- feb 02, 0.9992, 9, true (no page)
- jan 23, 1.0000, 12, false?
- jan 22, 0.9999, 6, true
- jan 22, 0.9997, 9, true
- jan 16, 0.9997, 9, true?
- jan 10, 1.0000, 13, true
- jan 07, 0.9995, 5, false?

Based on these numbers I don't feel comfortable changing the probably cut
offs (0.9999 for an alert, 0.9995 for a page).

But I do see an opertunity to raise the required number of tickets
for paging. Currently all alerts must have 5 tickets. This diff
raises the threshold to 7 tickets for paging. The effect on this
year's pages would be to remove 3 false pages, and exclude one true
one (which would still create an alert, simply not a page).

Test Plan: - Fingers crossed.

Reviewers: jacqueline

Subscribers: